# target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Software Defined NFV-assisted Wireless Network REST API","description":"Software Defined NFV-assisted Wireless Network REST API"},"basePath":"/onos/wifimaster","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"sdvrs","description":"Sample web resource"}],"paths":{"/sdvrs/wifi/removeVap":{"post":{"tags":["sdvrs"],"summary":"Post an exist virtual access point remove request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/RemoveVap"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/tenant/delTenant":{"post":{"tags":["sdvrs"],"parameters":[{"name":"stream","in":"body","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/wifi/removeNvap":{"post":{"tags":["sdvrs"],"summary":"Post a new virtual access point add request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/RemoveNvap"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/tenant/addTenant":{"post":{"tags":["sdvrs"],"summary":"Post an add tenant request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/addTenant"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/wifi/createNvap":{"post":{"tags":["sdvrs"],"summary":"Post a new virtual access point add request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/CreateNvap"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/vnf/createVnf":{"post":{"tags":["sdvrs"],"summary":"Post a new virtual network function request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/CreateVnf"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/wifi/createVap":{"post":{"tags":["sdvrs"],"summary":"Post a new virtual access point add request","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/CreateVap"},"description":"JSON stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/sdvrs/hello":{"get":{"tags":["sdvrs"],"summary":"Get hello world greeting","description":"","parameters":[],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{"CreateVap":{"type":"object","required":["tenantId","wifiAgent","interface","ssid"],"properties":{"tenantId":{"type":"integer","format":"int64","example":2},"wifiAgent":{"type":"String","example":"192.168.109.205"},"interface":{"type":"String","example":"wlan1"},"ssid":{"type":"String","example":"POF"}}},"CreateNvap":{"type":"object","required":["wifiAgent","vapNumber"],"properties":{"vapNumber":{"type":"String","example":"2"},"wifiAgent":{"type":"String","example":"192.168.109.205"}}},"RemoveNvap":{"type":"object","required":["wifiAgent","vapNumber"],"properties":{"vapNumber":{"type":"String","example":"2"},"wifiAgent":{"type":"String","example":"192.168.109.205"}}},"RemoveVap":{"type":"object","required":["tenantId","wifiAgent","interface"],"properties":{"tenantId":{"type":"integer","format":"int64","example":1},"wifiAgent":{"type":"String","example":"192.168.109.205"},"interface":{"type":"String","example":"wlan1"}}},"addTenant":{"type":"object","required":["tenantId","tenantName"],"properties":{"tenantId":{"type":"integer","format":"int64","example":1},"tenantName":{"type":"String","example":"Alice"}}},"CreateVnf":{"type":"object","required":["tenantId","vnfAgent","vnfId","vnfType","nbPort","script"],"properties":{"tenantId":{"type":"integer","format":"int64","example":1},"vnfAgent":{"type":"String","example":"192.168.109.205"},"vnfId":{"type":"integer","format":"int64","example":2},"vnfType":{"type":"String","example":"FireWall"},"nbPort":{"type":"integer","format":"int64","example":2},"script":{"type":"String","example":"iptables -A FORWARD -p tcp --dport 80 -j ACCEPT"}}}}}